+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add an optional `session_name` kwarg to the `Client` constructor to allow extra information to be incremented into the User-Agent header.
 - Add "region" parameter to CreateUser and CreateUserWithMutator to allow specifying in which region the user data should reside
 - Rename "compositeunion" and "compositeintersection" types of access policies to "composite_or" and "composite_and"
+- Breaking change: Added AsyncClient class and UCHttpAsyncClient interface to support asynchronous requests. As part of this, Error class was consolidated into UserCloudsSDKError, which may break any customers referencing it.
 
 ## 1.3.0 - 11-12-2023